requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
black==19.3b0
bleach==3.1.0
chardet==3.0.4
Click==7.0
dateparser==0.7.1
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
invoice2data==0.3.5
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.1
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
PyMuPDF==1.16.1
pyrsistent==0.15.4
pytesseract==0.3.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
regex==2019.8.19
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
tornado==6.0.3
traitlets==4.3.2
tzlocal==2.0.0
Unidecode==1.1.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
